refactored PhaserConfig exported class Level1
